% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_coexp_network.R
\name{build_coexp_direct_network}
\alias{build_coexp_direct_network}
\title{Build Direct Co-Expression network}
\usage{
build_coexp_direct_network(
  estimated_expression,
  genes,
  nlambda = 30,
  verbose = FALSE
)
}
\arguments{
\item{estimated_expression}{data.frame with columns
gene_id
run_accession}

\item{genes}{vector of gene labels}

\item{verbose}{verbose output (default: FALSE)}
}
\value{
result of huge::huge.select
    In particular .$merge[<nlambda>] is a n_gene by n_gene co-expression matrix
}
\description{
Use the graphical lasso to estimate a sparse co-expression matrix huge package
}
