% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_coexp_network.R
\name{build_coexp_intra_study_spearman_network}
\alias{build_coexp_intra_study_spearman_network}
\title{Build Co-exp network for each study and average}
\usage{
build_coexp_intra_study_spearman_network(
  estimated_expression,
  genes,
  verbose = FALSE
)
}
\arguments{
\item{estimated_expression}{data.frame with columns
study_accession,
run_accession,
gene_id,
FPKM}

\item{genes}{vector of gene labels}

\item{verbose}{verbose output (default: FALSE)}
}
\value{
n_gene by n_gene matrix with values with the average
    spearman rank correlation across studies of of the FPKM values
    for each run in the study
}
\description{
Build Co-exp network for each study and average
}
