% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_expression_data.R
\name{download_expression_data}
\alias{download_expression_data}
\title{Download RNA-Seq expression runs from SRA}
\usage{
download_expression_data(runs, sra_dir)
}
\arguments{
\item{runs:}{data.frame with column 'run_accession'}

\item{sra_dir:}{directory where to save the expression data}
}
\description{
Use prefetch to download data from the SRA to
}
\details{
<sra_dir>/<run_accesion>/<run_accession>.sra

  Each expression run is ~8 Gb, so roughly 1000 runs will require
  approximately 8 terabytes of space
}
