% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_expression.R
\name{estimate_expression}
\alias{estimate_expression}
\title{Estimate expression data from SRA run and reference genome}
\usage{
estimate_expression(
  run_accession,
  sra_fname,
  is_paired,
  results_dir,
  logs_dir,
  work_dir,
  reference_genome_path,
  fastq_dump_program,
  rsem_calculate_expression_program,
  bowtie2_path,
  n_threads = 1
)
}
\arguments{
\item{run_accession}{identifier for exprsesion run}

\item{sra_fname}{path to .sra file for the run}

\item{is_paired}{is it paired end read expression data?}

\item{results_dir}{path where the results should be stored}

\item{logs_dir}{path where the logs should be stored}

\item{work_dir}{path where intermediate data should be stored}

\item{reference_genome_path}{path to reference genome}

\item{fastq_dump_program}{path to fastq_dump}

\item{rsem_calculate_expression_program}{path to rsem-calculate-expression}
}
\description{
Example
}
\details{
tag <- "20201209"
estimate_expression(
   run_accession = "SRR513613",
   sra_fname = "/scratch/maom_root/maom99/maom/CryptoCEN/sra/SRR513613/SRR513613.sra",
   is_paired = TRUE,
   results_dir = "intermediate_data/estimated_expression_20201209",
   results_dir = "intermediate_data/estimated_expression_20201209/logs",
   work_dir = "/scratch/maom_root/maom99/maom/CryptoCEN/estimated_expression_20201209",
   reference_genome_path = "/scratch/maom_root/maom99/maom/CryptoCEN/reference_geneome",
   fastq_dump_program = "
}
