% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_estimated_expression.R
\name{gather_estimated_expression}
\alias{gather_estimated_expression}
\title{Gather estimated expression data for analysis}
\usage{
gather_estimated_expression(results_dir, verbose = TRUE)
}
\arguments{
\item{results_dir}{path used in e.g. estimate_expression(...)}

\item{verbose}{(default: true)}
}
\value{
a data.frame with columns (see RSEM for interpretation)
    run_accession
    gene_id
    transcript_ids
    length
    effective_length
    expected_count
    TPM
    FPKM
and one row per (run_accession, gene_id)
}
\description{
After estimating expression locally or in parallel
gather_estimated_expression parses files of the form
}
\details{
<results_dir>/<run_accession>.genes.results
}
