% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_estimated_expression.R
\name{gather_estimated_expression_meta}
\alias{gather_estimated_expression_meta}
\title{Gather estimated expression data metadata}
\usage{
gather_estimated_expression_meta(results_dir, verbose = TRUE)
}
\arguments{
\item{results_dir}{path used in e.g. estimate_expression(...)}

\item{verbose}{(default: true)}
}
\value{
tibble::tibble() with rows representing runs and columns
    n_reads
    n_reads_aligned_0_times
    n_reads_aligned_1_time
    n_reads_aligned_mutiple_times
    overall_alignment_percent
    run_time
}
\description{
Gather estimated expression data metadata
}
