% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_fri.R
\name{get_deepfri_predictions}
\alias{get_deepfri_predictions}
\title{Use the DeepFRI webserver to get sequence and structure based gene function prediction}
\usage{
get_deepfri_predictions(z, tag = NULL, output_path = NULL, verbose = TRUE)
}
\arguments{
\item{tag}{used to tracking the prediction on the DeepFRI webserver and for
the output path if not using use the basename without .pdb extension.
(default: NULL)}

\item{verbose}{use verbose output (default: TRUE)}

\item{pdb_fname}{proteins structure .pdb file}

\item{output_fname}{folder where the output predictions should be written
if null, then don't save output to file. (default: NULL)}
}
\value{
data.frame with columns
    structure_tag: <tag>
    prediction_type: [cnn, gcn] for sequence and structure based predictions
    go_term: e.g. GO:0044237
    go_term_name: e.g. "cellular metabolic process"
    go_termscore: prediction score
}
\description{
Using the https://beta.deepfri.flatironinstitute.org/ webserver
 Use Structure-based gene function predictions using graph convolutional networks
 (Gligorijevic, et al., 2019)
}
\details{
If you use for more than few predictions please coordinate with the Bonneau lab
 to not abuse their resources

 usage example:
   On disk: <tag>.pdb
   In R:

    predictions <- get_deepfri_predictions(
       pdb_fname = "<tag>.pdb",
       output_path = "intermediate_data")

If you use please cite:

	Gligorijevic, Vladimir and Renfrew, P. Douglas and Kosciolek, Tomasz and Leman,
     Julia Koehler and Cho, Kyunghyun and Vatanen, Tommi and Berenberg, Daniel
	and Taylor, Bryn and Fisk, Ian M. and Xavier, Ramnik J. and Knight, Rob and Bonneau, Richard
	Structure-Based Function Prediction using Graph Convolutional Networks
	2019, https://www.biorxiv.org/content/early/2019/10/04/786236
}
