% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_runs.R
\name{get_sra_connection}
\alias{get_sra_connection}
\title{Get connection to the sequence read archive database}
\usage{
get_sra_connection(parameters, verbose = TRUE)
}
\description{
if it doesn't exist download it to
}
\details{
<parameters$data_path$scratch_dir>/sra_meta/SRAmetadb.sqlite
}
