% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Load parameters}
\usage{
load_parameters(path = "parameters.yaml", verbose = TRUE)
}
\arguments{
\item{path}{Location to parameters.yaml file (default: parameters.yaml)}

\item{verbose}{(default: TRUE)}
}
\description{
Load parameters
}
