% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_gba.R
\name{network_gba}
\alias{network_gba}
\title{Evaluate network Guilt By Association and report AUROC scores}
\usage{
network_gba(
  genes,
  annotation_sets,
  networks,
  min = 20,
  max = 1000,
  nfold = 3,
  degrees = 1:length(networks),
  verbose = FALSE
)
}
\arguments{
\item{genes}{list of genes}

\item{annotation_sets}{list of go term annotations to predict}

\item{networks}{list of networks used for prediction}

\item{min}{filter terms requiring a minimum number of annotations (Default: 20)}

\item{max}{filter terms requiring a maximum number of annotations (Default: 1000)}

\item{nfold}{(Default 3)}

\item{degrees}{(Default 1:length(networks))}

\item{verbose}{(Default FALSE)}
}
\value{
data.frame with columns
   anno_id
   auroc_mean
   auroc_std
   degree_null_auroc_mean
   degree_null_auroc_std
}
\description{
Evaluate network Guilt By Association and report AUROC scores
}
