% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_heatmap.R
\name{plot_network_heatmap}
\alias{plot_network_heatmap}
\title{Plot network heatmap}
\usage{
plot_network_heatmap(
  network,
  output_fname,
  gene_order = "TSP",
  verbose = FALSE
)
}
\arguments{
\item{network}{matrix}

\item{output_fname}{(should be .png)}

\item{gene_order}{sort the genes using the traveling salesman path using the seriation package}

\item{verbose}{(Default: FALSE)}
}
\description{
Create a grayscale heatmap for large matrix
}
