% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reference_genome.R
\name{prepare_reference_genome}
\alias{prepare_reference_genome}
\title{Use RSEM to prepare reference genome}
\usage{
prepare_reference_genome(parameters, verbose = TRUE)
}
\arguments{
\item{parameters}{as read in by load_parameters()}

\item{verbose}{(default: TRUE)}
}
\description{
use rsem-prepare-reference to generate reference genome
for mapping RNAseq transcripts
}
\details{
# Usage example:
in parameters.yaml

  source_data:
    genome:
       transcript_fasta_path: "raw_data/FungiDB-49_CneoformansH99_AnnotatedTranscripts.fasta"
  data_paths:
    reference_genome_path: "/scratch/maom_root/maom99/maom/CryptoCEN/reference_genome/CneoformansH99"
  software_paths:
    bowtie2_path: "/home/maom/opt/bowtie2-2.3.4.1-linux-x86_64"
    fastq_dump_program: "/home/maom/opt/bin/fastq-dump"
    rsem_path: "/home/maom/opt/RSEM/"

In R:

  library(CalCEN)
  parameters <- CalCEN::load_parameters()
  CalCEN::prepare_reference_genome()
}
