% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_annotations.R
\name{propagate_go_annotations}
\alias{propagate_go_annotations}
\title{Propagate GO term annotations}
\usage{
propagate_go_annotations(go_annotations, verbose = TRUE)
}
\arguments{
\item{go_annotations}{data.frame with column go_id and object identifiers}

\item{verbose}{(Default: TRUE)}
}
\value{
data.frame with columns of input <go_annotations>
  where the annotations are propagated up the hierarchy
}
\description{
GO terms are organized into a directed acyclic graph where an
annotation fo a more specific term implies the annotation of the
more general.
}
\details{
GO terms are not
}
