% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_runs.R
\name{retrieve_sra_runs}
\alias{retrieve_sra_runs}
\title{Retrieve SRA runs from the SRAmeta database}
\usage{
retrieve_sra_runs(sra_con, curated_studies, verbose = TRUE)
}
\arguments{
\item{sra_con}{a DBI connection to the SRA database. E.g. the
result of calling get_sra_connection()}

\item{curated_studies}{a character vector with study_accession values
or a data.frame with a column study_accession}

\item{verbose}{(default: TRUE)}
}
\value{
data.frame with one run per row with columns in
    curated_studies an additionally:

        study_alias
        study_title
        experiment_accession
        sample_accession
        library_strategy
        library_layout
        sample_alias
        taxon_id
        scientific_name
        common_name
        run_accession
        run_alias
}
\description{
Retrieve SRA runs from the SRAmeta database
}
