% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_expression_data.R
\name{validate_downloaded_runs}
\alias{validate_downloaded_runs}
\title{Check that the downloaded runs are not corrupted}
\usage{
validate_downloaded_runs(sra_dir)
}
\arguments{
\item{sra_dir:}{path to find downloaded SRA runs}
}
\description{
use vdb-validate from the prefetch package to validate all
   all .sra files under
}
\details{
<sra_dir>/<run_accession>/<run_accession>.sra
}
