% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_expression.R
\name{validate_estimated_runs}
\alias{validate_estimated_runs}
\title{Validate estimated expression runs}
\usage{
validate_estimated_runs(path, check_for_logs = TRUE)
}
\arguments{
\item{check_for_logs}{check for logs}

\item{results_dir}{directory to look for completed estimated expression runs}
}
\value{
data.frame with column [run_accession] of runs that have been completed
}
\description{
After running estimate_expression(..., results_dir = <results_dir>, ...)
call validate_estimated_runs(results_dir = <results_dir>)
to get the runs that have been completed
}
