% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_runs.R
\name{get_sra_connection}
\alias{get_sra_connection}
\title{Get connection to the sequence read archive database}
\usage{
get_sra_connection(parameters, verbose = TRUE)
}
\arguments{
\item{parameters}{list having element data_paths$scratch_dir}

\item{verbose}{print verbose output}
}
\value{
dtabase connection to the sra metadata database
}
\description{
if it doesn't exist download it to

  <parameters$data_path$scratch_dir>/sra_meta/SRAmetadb.sqlite
}
